+++ tests/V1/Terminals/Meta/T_FRACTIONAL_PERCENTAGE_Test.php
@@ -59,7 +59,7 @@ class T_FRACTIONAL_PERCENTAGE_Test extends BaseTestCase
 
     protected function getExpectedPseudoBNF()
     {
-        return 'regex /^(1\\.0|0\\.[0-9]+|1|0)(?![0-9%])/';
+        return 'regex /^(1\\.0|0\\.[0-9]+|1|0)(?![0-9a-fA-F%])/';
     }
 
     protected function getDatasetKeysToMatch()