+++ tests/V1/Terminals/Strict/T_INT_1_Test.php
@@ -59,7 +59,7 @@ class T_INT_1_Test extends BaseTestCase
 
     protected function getExpectedPseudoBNF()
     {
-        return 'regex /^1(?![0-9\\.%])/';
+        return 'regex /^1(?![0-9a-fA-F\\.%])/';
     }
 
     protected function getDatasetKeysToMatch()